Broken down the Asset Details screen in smaller components
